requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.44
alembic==1.13.2
pydantic==2.9.2
python-multipart==0.0.10
cryptography==43.0.3
qrcode==7.4.2
Pillow==10.3.0
python-dotenv==1.0.1